Add xgmii_fifo on the receive and transmit paths.
The tx fifo might not be neccesary but might add a bit more
timing slack.

Also include constraints from verilog-ethernet and try to
fix other constraints.  Resort to a sledghammer saying that
the axi_clock is not related to any other clocks which
fixes a lot of the timing problems from before.